requirements.txt
llama-cpp-python
duckduckgo_search
colorama
requests
beautifulsoup4
trafilatura
readchar
